adding except option to cleanup, checkCleanup moved to its own file
